Move deeplink parser implementation to its own file
